+++ packages/core/src/typings/args/getAssetTransfersArgs.ts
@@ -1,0 +1,13 @@
+
+/**
+ * The argument object for [[AssetApi.getAssetTransfers]]
+ *
+ * @module core
+ */
+export interface GetAssetTransfersArgs {
+    assetId?: string;
+    accountId?: string;
+    includeAssetInfo?: boolean;
+    firstIndex?: number;
+    lastIndex?: number;
+}