+++ packages/core/src/typings/args/getAssetTransfersPerAccountArgs.ts
@@ -1,0 +1,12 @@
+
+/**
+ * The argument object for [[AssetApi.getAssetTransfersPerAccount]]
+ *
+ * @module core
+ */
+export interface GetAssetTransfersPerAccountArgs {
+    accountId: string;
+    includeAssetInfo?: boolean;
+    firstIndex?: number;
+    lastIndex?: number;
+}