+++ packages/core/src/typings/args/getAssetTransfersPerAssetArgs.ts
@@ -1,0 +1,12 @@
+
+/**
+ * The argument object for [[AssetApi.getAssetTransfersPerAsset]]
+ *
+ * @module core
+ */
+export interface GetAssetTransfersPerAssetArgs {
+    assetId: string;
+    includeAssetInfo?: boolean;
+    firstIndex?: number;
+    lastIndex?: number;
+}