+++ packages/core/src/typings/assetTransferList.ts
@@ -1,0 +1,14 @@
+/**
+ * Original work Copyright (c) 2022 Signum Network
+ */
+
+import {AssetTransfer} from './assetTransfer';
+
+/**
+ * Asset Transfer List
+ ** @module core
+ */
+export interface AssetTransferList {
+    transfers: AssetTransfer[];
+    requestProcessingTime: number;
+}